ModbusTCP_sensor/flask_app/requirements.txt
flask==2.3.3
pymodbus==3.7.2
psycopg2-binary==2.9.10
